requirements.txt
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-experimental
pandas
numpy
python-dotenv==1.0.1
tabulate